Revert "TX: Support ThreadPoolManager to manage metrics and re-config of threadPool"
This reverts commit 8c81a0059c321157b739c72d848d674a10cbc08c.

+++ server/src/main/java/org/apache/uniffle/server/buffer/ShuffleBufferManager.java
@@ -146,11 +146,12 @@ public class ShuffleBufferManager {
     ShuffleServerMetrics.addLabeledCacheGauge(
         BLOCK_COUNT,
         () ->
-            bufferPool.values().stream()
-                .flatMap(innerMap -> innerMap.values().stream())
-                .flatMap(rangeMap -> rangeMap.asMapOfRanges().values().stream())
-                .mapToInt(shuffleBuffer -> shuffleBuffer.getBlockCount())
-                .sum(),
+            (double)
+                bufferPool.values().stream()
+                    .flatMap(innerMap -> innerMap.values().stream())
+                    .flatMap(rangeMap -> rangeMap.asMapOfRanges().values().stream())
+                    .mapToInt(shuffleBuffer -> shuffleBuffer.getBlockCount())
+                    .sum(),
         2 * 60 * 1000L /* 2 minutes */);
   }
